pyproject.toml
[tool.poetry]
name = "physicalyb"
version = "0.1.0"
description = ""
authors = ["akash-banger2 <akashbanger@troopapp.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
django-rest-framework = "^0.1.0"
cors = "^1.0.1"
django-cors-headers = "^4.2.0"
requests = "^2.31.0"
pillow = "^10.0.0"
numpy = "^1.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
